+++ CHANGELOG.md
@@ -11,6 +11,13 @@ All notable changes to this project will be documented in this file.
 - New top-level config field `shutdown_delay` for delaying graceful termination.
 - New `snowflake_id` bloblang function.
 - Field `wait_time_seconds` added to the `aws_sqs` input.
+- New `json_path` bloblang method.
+
+### Fixed
+
+- The `schema_registry_encode` processor now correctly assumes Avro JSON encoded documents by default.
+- The `redis` processor `retry_period` no longer shows linting errors for duration strings.
+- The `/inputs` and `/outputs` endpoints for dynamic inputs and outputs now correctly render configs, both structured within the JSON response and the raw config string.
 
 ## 4.5.1 - 2022-08-10